Micro optimisations in calculateCanTouchForType
This function is called /very/ often. #players * #entities * #touch types to
be precise. Any speedup in this function will magnify by that factor, which
made it worth the effort. Reducing the source of the complexity requires way
more effort than I want to spend on it, so this optimisation is one of the
next best things.

In FProfiler I've seen the average time of this function go down by about a
third, but since the times are in the 0.0x range I'm not sure how reliable
that is.

See #208